requirements.txt
speedtest-cli
requests
